pyproject.toml
[project]
name = "borgboi"
version = "1.17.3"
description = "Add your description here"
license = "MIT"
readme = "README.md"
authors = [{ name = "Zach Fuller", email = "zach.fuller1222@gmail.com" }]
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.35.92",
    "catppuccin[pygments]>=2.4.1",
    "cyclopts>=4.8.0",
    "pydantic>=2.10.4",
    "pydantic-settings>=2.12.0",
    "pyyaml>=6.0",
    "rich>=13.9.4",
    "sqlalchemy>=2.0",
]

[project.scripts]
borgboi = "borgboi.cli:main"
bb = "borgboi.cli:main"

[project.urls]
Homepage = "https://github.com/fullerzz/borgboi"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "basedpyright>=1.29.1",
    "boto3-stubs[essential]>=1.35.92",
    "commitizen>=4.8.3",
    "git-cliff>=2.10.1",
    "inline-snapshot>=0.32.3",
    "moto[dynamodb,iam,s3]>=5.0.26",
    "mypy[faster-cache]>=1.14.1",
    "prek>=0.3.2",
    "pytest>=8.3.4",
    "pytest-clarity>=1.0.1",
    "pytest-cov>=6.0.0",
    "pytest-socket>=0.7.0",
    "pytest-sugar>=1.0.0",
    "ruff>=0.8.6",
    "rust-just>=1.39.0",
    "ty>=0.0.1a12",
    "types-pyyaml>=6.0.12",
]
docs = [
    "zensical>=0.0.24",
]

[tool.uv]
default-groups = ["dev", "docs"]

[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
    "--disable-socket",
    "--allow-unix-socket",
    "--cov=src/borgboi",
    "--cov-report=term",
]
filterwarnings = [
    "error",
    "ignore::DeprecationWarning:botocore.auth:422", # handles: 'DeprecationWarning: datetime.datetime.utcnow()'
    "ignore::DeprecationWarning:botocore.auth:424", # handles: 'DeprecationWarning: datetime.datetime.utcnow()'
    "ignore::DeprecationWarning:botocore.auth:425", # handles: 'DeprecationWarning: datetime.datetime.utcnow()'
]


[tool.pyright]
include = ["src/**", "tests/**"]
typeCheckingMode = "standard"

[tool.mypy]
plugins = ["pydantic.mypy"]
warn_redundant_casts = true
warn_unused_ignores = true
disallow_any_generics = true
check_untyped_defs = true
no_implicit_reexport = true
disallow_untyped_defs = true
local_partial_types = true
warn_return_any = true
enable_error_code = [
    "ignore-without-code",
    "no-untyped-def",
    "unreachable",
    "unused-awaitable",
    "explicit-override",
    "mutable-override",
    "exhaustive-match",
]

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

[tool.ruff]
include = ["src/**", "tests/**"]
exclude = ["tests/data/*.txt"]
line-length = 120
indent-width = 4

target-version = "py313"

[tool.ruff.lint]
select = [
    "E",
    "F",
    "W",
    "C90",
    "I",
    "N",
    "UP",
    "ASYNC",
    "S",
    "B",
    "ERA",
    "PLE",
    "PLW",
    "PLC",
    "PLW",
    "PERF",
    "RUF",
    "SIM",
    "PT",
    "T20",
    "PTH",
    "LOG",
    "G",
]
ignore = ["E501", "S101", "PLC0415"]

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.inline-snapshot]
format-command = "ruff format --stdin-filename {filename}"

[tool.ty.rules]
possibly-unresolved-reference = "warn"

[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "v$major.$minor.$patch$prerelease"
version_scheme = "pep440"
version_provider = "pep621"
update_changelog_on_bump = false
major_version_zero = false
bump_message = "bump: version $current_version -> $new_version"
changelog_file = "CHANGELOG.md"
